refactor(YouTube - Download Playlist Button): Aggregate download actions within the same patch
