+++ src/main/kotlin/app/revanced/patches/youtube/player/overlaybuttons/OverlayButtonsBytecodePatch.kt
@@ -1,22 +1,19 @@
 package app.revanced.patches.youtube.player.overlaybuttons
 
 import app.revanced.patcher.data.BytecodeContext
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructionsWithLabels
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.patch.annotation.Patch
 import app.revanced.patcher.util.smali.ExternalLabel
-import app.revanced.patches.youtube.player.overlaybuttons.fingerprints.*
-import app.revanced.patches.youtube.utils.integrations.Constants.INTEGRATIONS_PATH
+import app.revanced.patches.youtube.player.overlaybuttons.fingerprints.PlayerButtonConstructorFingerprint
 import app.revanced.patches.youtube.utils.integrations.Constants.UTILS_PATH
 import app.revanced.patches.youtube.utils.mainactivity.MainActivityResolvePatch
 import app.revanced.patches.youtube.utils.resourceid.SharedResourceIdPatch
 import app.revanced.patches.youtube.video.information.VideoInformationPatch
 import app.revanced.util.*
 import com.android.tools.smali.dexlib2.Opcode
-import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.TwoRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.reference.FieldReference
@@ -31,49 +28,13 @@ import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 )
 object OverlayButtonsBytecodePatch : BytecodePatch(
     setOf(
-        OfflineVideoEndpointFingerprint,
-        PiPPlaybackFingerprint,
-        PlayerButtonConstructorFingerprint,
-        AccessibilityOfflineButtonSyncFingerprint,
-        HookDownloadPlaylistButtonFingerprint
+        PlayerButtonConstructorFingerprint
     )
 ) {
     private const val INTEGRATIONS_ALWAYS_REPEAT_CLASS_DESCRIPTOR =
         "$UTILS_PATH/AlwaysRepeatPatch;"
 
-    private const val INTEGRATIONS_VIDEO_UTILS_CLASS_DESCRIPTOR =
-        "$INTEGRATIONS_PATH/utils/VideoUtils;"
-
     override fun execute(context: BytecodeContext) {
-
-        // region patch for hook download button
-
-        OfflineVideoEndpointFingerprint.resultOrThrow().mutableMethod.apply {
-            addInstructionsWithLabels(
-                0, """
-                    invoke-static/range {p3 .. p3}, $INTEGRATIONS_VIDEO_UTILS_CLASS_DESCRIPTOR->inAppDownloadButtonOnClick(Ljava/lang/String;)Z
-                    move-result v0
-                    if-eqz v0, :show_native_downloader
-                    return-void
-                    """, ExternalLabel("show_native_downloader", getInstruction(0))
-            )
-        }
-
-        PiPPlaybackFingerprint.resultOrThrow().let {
-            it.mutableMethod.apply {
-                val insertIndex = it.scanResult.patternScanResult!!.endIndex
-                val insertRegister = getInstruction<OneRegisterInstruction>(insertIndex).registerA
-
-                addInstructions(
-                    insertIndex, """
-                        invoke-static {v$insertRegister}, $INTEGRATIONS_VIDEO_UTILS_CLASS_DESCRIPTOR->getExternalDownloaderLaunchedState(Z)Z
-                        move-result v$insertRegister
-                        """
-                )
-            }
-        }
-
-        // endregion
 
         // region patch for always repeat and pause
